hypervisor: change VcpuX86_64::interrupt to only accept u8 vector
Under the hood, both AMD [1] and Intel [2] only accept 8 bit vector for event injection, so it's highly likely that all hypervisor implementation to encounter an error when we pass in values greater than 0xff as the vector.

From the source we can tell that KVM rejects any value greater than 0xff, HAXM has some strange unwanted behaviors with invalid vector values. From testing, we can tell WHPX also rejects invalid vector values right away.

[1]: Figure 15-5. EVENTINJ Field in the VMCB of AMD64 Architecture Programmer’s Manual, Volumes 2.
[2]: Table 25-17. Format of the VM-Entry Interruption-Information Field of Intel 64 and IA-32 Architectures Software Developer's Manual Volume 3.

TEST=./tools/dev_container cargo nextest run --workspace --profile=default -E 'package(hypervisor) & binary(hypervisor_virtualization)' --retries 0 --no-capture
BUG=b:334055761

Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5602270
Commit-Queue: Kaiyi Li <kaiyili@google.com>

+++ devices/src/irqchip/userspace.rs
@@ -566,7 +566,7 @@ impl<V: VcpuX86_64 + 'static> IrqChip for UserspaceIrqChip<V> {
             let mut pic = self.pic.lock();
             if vcpu_ready {
                 if let Some(vector) = pic.get_external_interrupt() {
-                    vcpu.interrupt(vector as u32)?;
+                    vcpu.interrupt(vector)?;
                     self.apics[vcpu_id].lock().set_mp_state(&MPState::Runnable);
                     // Already injected a PIC interrupt, so APIC fixed interrupt can't be injected.
                     vcpu_ready = false;
@@ -595,7 +595,7 @@ impl<V: VcpuX86_64 + 'static> IrqChip for UserspaceIrqChip<V> {
             };
 
             if do_interrupt {
-                vcpu.interrupt(vector as u32)?;
+                vcpu.interrupt(vector)?;
             }
         }
         for _ in 0..irqs.nmis {
